+++ changelog.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+* Add a page to show an overview count by build environment and unique ports
+
 Version 1.7.3
 -------------